Move sphinx.directives.other:Index to sphinx.domains.index
